Api/post

+++ SocialMediaAppAPI/Controllers/PostsController.cs
@@ -7,12 +7,14 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using SocialMediaAppAPI.Data;
 using SocialMediaAppAPI.Models;
+using SocialMediaAppAPI.Types.Attributes;
 using SocialMediaAppAPI.Types.Requests;
 
 namespace SocialMediaAppAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [ValidateApiToken]
     public class PostsController : ControllerBase
     {
         private readonly APIDbContext _context;
@@ -22,13 +24,36 @@ namespace SocialMediaAppAPI.Controllers
             _context = context;
         }
 
+        private User GetAuthenticatedUser()
+        {
+            return HttpContext.Items["AuthenticatedUser"] as User;
+        }
+
         // GET: api/Posts/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<IEnumerable<Post>>> GetPost(Guid id)
+        public async Task<ActionResult<PostDTO>> GetPost(Guid id)
         {
+            var authenticatedUser = GetAuthenticatedUser();
+            if (authenticatedUser == null)
+            {
+                return Unauthorized();
+            }
+
             var post = await _context.Posts
-                                        .Where(c => c.Id == id)
-                                        .FirstOrDefaultAsync();
+                .Where(c => c.Id == id)
+                .Select(post => new PostDTO
+                {
+                    Id = post.Id,
+                    Content = post.Content,
+                    OPUsername = _context.Users.Where(u => u.Id == post.UserId).Select(u => u.UserName).FirstOrDefault(),
+                    Following = _context.Followers.Any(f => f.UserId == authenticatedUser.Id && f.FollowedUserId == post.UserId),
+                    LikeCount = post.LikeCount,
+                    CommentCount = post.CommentCount,
+                    CreatedAt = post.CreatedAt,
+                    UserId = post.UserId,
+                    Attachments = post.Attachments
+                })
+                .FirstOrDefaultAsync();
 
             if (post == null)
             {
@@ -38,10 +63,16 @@ namespace SocialMediaAppAPI.Controllers
             return Ok(post);
         }
 
-        // GET: api/Feed/{userId}/{page}/{amount}
-        [HttpGet("/api/Feed/Following/{userId}/{page}/{amount}")]
-        public async Task<ActionResult<IEnumerable<PostDTO>>> GetFollowingFeed(Guid userId, int page, int amount)
+        // GET: api/Feed/Following/{page}/{amount}
+        [HttpGet("/api/Feed/Following/{page}/{amount}")]
+        public async Task<ActionResult<IEnumerable<PostDTO>>> GetFollowingFeed(int page, int amount)
         {
+            var authenticatedUser = GetAuthenticatedUser();
+            if (authenticatedUser == null)
+            {
+                return Unauthorized();
+            }
+
             if (page == 0)
                 page = 1;
 
@@ -50,52 +81,44 @@ namespace SocialMediaAppAPI.Controllers
 
             var skip = (page - 1) * amount;
 
-            // Get the list of users the current user is following
             var followedUserIds = await _context.Followers
-                       .Where(f => f.UserId == userId)
-                       .Select(f => f.FollowedUserId)
-                       .ToListAsync();
+                .Where(f => f.UserId == authenticatedUser.Id)
+                .Select(f => f.FollowedUserId)
+                .ToListAsync();
 
-
-
-
-            // Get posts from followed users
             var followedPosts = _context.Posts
-                                         .Where(p => followedUserIds.Contains(p.UserId))
-                                         .OrderByDescending(p => p.CreatedAt);
+                .Where(p => followedUserIds.Contains(p.UserId))
+                .OrderByDescending(p => p.CreatedAt);
 
-            // Get posts from non-followed users
-            //var randomPosts = _context.Posts
-            //                          .Where(p => !followedUserIds.Contains(p.UserId))
-            //                          .OrderByDescending(p => p.CreatedAt);
-
-            // Combine followed and random posts
-            //var combinedPosts = followedPosts.Concat(randomPosts);
-
-            // Apply pagination
             var paginatedPosts = await followedPosts
-                                         .Skip(skip)
-                                         .Take(amount)
-                                         .ToListAsync();
+                .Skip(skip)
+                .Take(amount)
+                .Select(post => new PostDTO
+                {
+                    Id = post.Id,
+                    Content = post.Content,
+                    OPUsername = _context.Users.Where(u => u.Id == post.UserId).Select(u => u.UserName).FirstOrDefault(),
+                    Following = followedUserIds.Contains(post.UserId),
+                    LikeCount = post.LikeCount,
+                    CommentCount = post.CommentCount,
+                    CreatedAt = post.CreatedAt,
+                    UserId = post.UserId,
+                    Attachments = post.Attachments
+                })
+                .ToListAsync();
 
-            // Map to PostDTO
-            var postDTOs = paginatedPosts.Select(post => new PostDTO
-            {
-                Id = post.Id,
-                Content = post.Content,
-                LikeCount = post.LikeCount,
-                CommentCount = post.CommentCount,
-                CreatedAt = post.CreatedAt,
-                UserId = post.UserId,
-                Attachments = post.Attachments
-            }).ToList();
-
-            return Ok(postDTOs);
+            return Ok(paginatedPosts);
         }
 
-        [HttpGet("/api/Feed/NotFollowing/{userId}/{page}/{amount}")]
-        public async Task<ActionResult<IEnumerable<PostDTO>>> GetNotFollowingFeed(Guid userId, int page, int amount)
+        [HttpGet("/api/Feed/NotFollowing/{page}/{amount}")]
+        public async Task<ActionResult<IEnumerable<PostDTO>>> GetNotFollowingFeed(int page, int amount)
         {
+            var authenticatedUser = GetAuthenticatedUser();
+            if (authenticatedUser == null)
+            {
+                return Unauthorized();
+            }
+
             if (page == 0)
                 page = 1;
 
@@ -104,42 +127,46 @@ namespace SocialMediaAppAPI.Controllers
 
             var skip = (page - 1) * amount;
 
-            // Get the list of users the current user is following
             var followedUserIds = await _context.Followers
-                       .Where(f => f.UserId == userId)
-                       .Select(f => f.FollowedUserId)
-                       .ToListAsync();
+                .Where(f => f.UserId == authenticatedUser.Id)
+                .Select(f => f.FollowedUserId)
+                .ToListAsync();
 
-            // Get posts from non-followed users
             var notFollowedPosts = _context.Posts
-                                            .Where(p => !followedUserIds.Contains(p.UserId) && p.UserId != userId)
-                                            .OrderByDescending(p => p.CreatedAt);
+                .Where(p => !followedUserIds.Contains(p.UserId) && p.UserId != authenticatedUser.Id)
+                .OrderByDescending(p => p.CreatedAt);
 
-            // Apply pagination
             var paginatedPosts = await notFollowedPosts
-                                         .Skip(skip)
-                                         .Take(amount)
-                                         .ToListAsync();
+                .Skip(skip)
+                .Take(amount)
+                .Select(post => new PostDTO
+                {
+                    Id = post.Id,
+                    Content = post.Content,
+                    OPUsername = _context.Users.Where(u => u.Id == post.UserId).Select(u => u.UserName).FirstOrDefault(),
+                    Following = followedUserIds.Contains(post.UserId),
+                    LikeCount = post.LikeCount,
+                    CommentCount = post.CommentCount,
+                    CreatedAt = post.CreatedAt,
+                    UserId = post.UserId,
+                    Attachments = post.Attachments
+                })
+                .ToListAsync();
 
-            // Map to PostDTO
-            var postDTOs = paginatedPosts.Select(post => new PostDTO
-            {
-                Id = post.Id,
-                Content = post.Content,
-                LikeCount = post.LikeCount,
-                CommentCount = post.CommentCount,
-                CreatedAt = post.CreatedAt,
-                UserId = post.UserId,
-                Attachments = post.Attachments
-            }).ToList();
-
-            return Ok(postDTOs);
+            return Ok(paginatedPosts);
         }
 
         // GET: api/Posts/1/10
         [HttpGet("{page}/{amount}")]
         public async Task<ActionResult<IEnumerable<PostDTO>>> GetPosts(int page, int amount)
         {
+            var authenticatedUser = HttpContext.Items["AuthenticatedUser"] as User;
+            if (authenticatedUser == null)
+            {
+                return BadRequest();
+                return Unauthorized();
+            }
+
             if (page == 0)
                 page = 1;
 
@@ -148,37 +175,47 @@ namespace SocialMediaAppAPI.Controllers
 
             var skip = (page - 1) * amount;
 
-            return await _context.Posts.Select(Post => new PostDTO
-            {
-                Id = Post.Id,
-                Content = Post.Content,
-                LikeCount = Post.LikeCount,
-                CommentCount = Post.CommentCount,
-                CreatedAt = Post.CreatedAt,
-                UserId = Post.UserId,
-                Attachments = Post.Attachments
-            })
-            .Skip(skip)
-            .Take(amount)
-            .ToListAsync();
-        }
+            var posts = await _context.Posts
+                .Skip(skip)
+                .Take(amount)
+                .Select(post => new PostDTO
+                {
+                    Id = post.Id,
+                    Content = post.Content,
+                    OPUsername = _context.Users.Where(u => u.Id == post.UserId).Select(u => u.UserName).FirstOrDefault(),
+                    Following = _context.Followers.Any(f => f.UserId == authenticatedUser.Id && f.FollowedUserId == post.UserId),
+                    LikeCount = post.LikeCount,
+                    CommentCount = post.CommentCount,
+                    CreatedAt = post.CreatedAt,
+                    UserId = post.UserId,
+                    Attachments = post.Attachments
+                })
+                .OrderByDescending(p => p.CreatedAt)
+                .ToListAsync();
 
+            return Ok(posts);
+        }
 
         // PUT: api/Posts/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
         public async Task<IActionResult> PutPost(Guid id, UpdatePostDTO post)
         {
+            var authenticatedUser = GetAuthenticatedUser();
+            if (authenticatedUser == null)
+            {
+                return Unauthorized();
+            }
+
             var existingPost = await _context.Posts
-                                        .Where(c => c.Id == id)
-                                        .FirstOrDefaultAsync();
+                .Where(c => c.Id == id)
+                .FirstOrDefaultAsync();
 
             if (existingPost == null)
             {
                 return NotFound();
             }
 
-            // Only allow updating the content of the comment, not the UserId
             existingPost.Content = post.Content;
 
             try
@@ -198,10 +235,16 @@ namespace SocialMediaAppAPI.Controllers
         [HttpPost]
         public async Task<ActionResult<Post>> PostPost(CreatePostDTO post)
         {
+            var authenticatedUser = GetAuthenticatedUser();
+            if (authenticatedUser == null)
+            {
+                return Unauthorized();
+            }
+
             var createdPost = new Post
             {
                 Id = Guid.NewGuid(),
-                UserId = post.UserId,
+                UserId = authenticatedUser.Id,
                 Content = post.Content,
                 LikeCount = 0,
                 CommentCount = 0,
@@ -211,21 +254,30 @@ namespace SocialMediaAppAPI.Controllers
             _context.Posts.Add(createdPost);
             await _context.SaveChangesAsync();
 
-            // Assuming you have an action called "GetPost" that takes id as a parameter
             return CreatedAtAction("GetPost", new { id = createdPost.Id }, createdPost);
         }
-
 
         // DELETE: api/Posts/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeletePost(Guid id)
         {
+            var authenticatedUser = GetAuthenticatedUser();
+            if (authenticatedUser == null)
+            {
+                return Unauthorized();
+            }
+
             var post = await _context.Posts
-                                        .Where(c => c.Id == id)
-                                        .FirstOrDefaultAsync();
+                .Where(c => c.Id == id)
+                .FirstOrDefaultAsync();
             if (post == null)
             {
                 return NotFound();
+            }
+
+            if (post.UserId != authenticatedUser.Id)
+            {
+                return Forbid();
             }
 
             _context.Posts.Remove(post);
